reorganize DTO folder
