+++ src/brands/brands.controller.ts
@@ -13,10 +13,10 @@ import {
 } from '@nestjs/common';
 import { FileInterceptor } from '@nestjs/platform-express';
 import { NonEmptyBodyPipe } from '@/common/pipes/non-empty-body.pipe';
-import { PaginationQueryDto } from '@/common/dtos/paginations.dto';
+import { PaginationQueryDto } from '@/common/dto/paginations.dto';
 import { BrandsService } from './brands.service';
-import { CreateBrandDto } from './dtos/create-brand.dto';
-import { UpdateBrandDto } from './dtos/update-brand.dto';
+import { CreateBrandDto } from './dto/create-brand.dto';
+import { UpdateBrandDto } from './dto/update-brand.dto';
 
 @Controller('brands')
 export class BrandsController {
