+++ src/categories/categories.controller.ts
@@ -13,10 +13,10 @@ import {
 } from '@nestjs/common';
 import { FileInterceptor } from '@nestjs/platform-express';
 import { NonEmptyBodyPipe } from '@/common/pipes/non-empty-body.pipe';
-import { PaginationQueryDto } from '@/common/dtos/paginations.dto';
+import { PaginationQueryDto } from '@/common/dto/paginations.dto';
 import { CategoriesService } from './categories.service';
-import { CreateCategoryDto } from './dtos/create-category.dto';
-import { UpdateCategoryDto } from './dtos/update-category.dto';
+import { CreateCategoryDto } from './dto/create-category.dto';
+import { UpdateCategoryDto } from './dto/update-category.dto';
 
 @Controller('categories')
 export class CategoriesController {
