+++ src/media-assets/dto/create-media-asset.dto.ts
